refactor(agent): split project tour native operations
